+++ flecsi/library.cmake
@@ -2,6 +2,9 @@ set(flecsi_HEADERS
   data.hh
   execution.hh
   flog.hh
+  runtime.hh
+  topology.hh
+  utilities.hh
 )
 
 set(flecsi_SOURCES