+++ flecsi/util/unit/main.cc
@@ -1,4 +1,4 @@
-#include <flecsi/execution.hh>
+#include <flecsi/runtime.hh>
 #include <flecsi/util/unit.hh>
 
 using namespace flecsi;