+++ flecsi/utilities.hh
@@ -1,0 +1,41 @@
+// Copyright (C) 2016, Triad National Security, LLC
+// All rights reserved.
+
+#ifndef FLECSI_UTILITIES_HH
+#define FLECSI_UTILITIES_HH
+
+/// \defgroup utils Utilities
+/// Non-domain-specific infrastructure.
+/// \code#include "flecsi/utilities.hh"\endcode
+///
+/// This header provides the features in the following additional \b deprecated
+/// headers:
+/// - \ref annotation "flecsi/util/annotation.hh"
+/// - \ref ranges "flecsi/util/array_ref.hh"
+/// - flecsi/util/color_map.hh
+/// - flecsi/util/common.hh
+/// - flecsi/util/dag.hh
+/// - flecsi/util/demangle.hh
+/// - \ref point "flecsi/util/dimensioned_array.hh"
+/// - \ref mpi-utils "flecsi/util/mpi.hh"
+/// - flecsi/util/reorder.hh
+/// - \ref serial "flecsi/util/serialize.hh"
+/// - flecsi/util/set_intersection.hh
+/// - flecsi/util/set_utils.hh
+/// - \ref unit "flecsi/util/unit.hh"
+
+#include "flecsi/util/annotation.hh"
+#include "flecsi/util/array_ref.hh"
+#include "flecsi/util/color_map.hh"
+#include "flecsi/util/common.hh"
+#include "flecsi/util/dag.hh"
+#include "flecsi/util/demangle.hh"
+#include "flecsi/util/dimensioned_array.hh"
+#include "flecsi/util/mpi.hh"
+#include "flecsi/util/reorder.hh"
+#include "flecsi/util/serialize.hh"
+#include "flecsi/util/set_intersection.hh"
+#include "flecsi/util/set_utils.hh"
+#include "flecsi/util/unit.hh"
+
+#endif