+++ tutorial/1-runtime/3-flog.cc
@@ -1,6 +1,6 @@
 #include <flecsi/execution.hh>
 #include <flecsi/flog.hh>
-#include <flecsi/run/control.hh>
+#include <flecsi/runtime.hh>
 
 #include <fstream>