+++ tutorial/2-control/3-dependencies.hh
@@ -2,7 +2,7 @@
 #define TUTORIAL_2_CONTROL_3_DEPENDENCIES_HH
 
 #include "flecsi/flog.hh"
-#include "flecsi/run/control.hh"
+#include "flecsi/runtime.hh"
 
 namespace dependencies {