+++ tutorial/4-data/control.hh
@@ -2,7 +2,7 @@
 #define CONTROL_HH
 
 #include <flecsi/flog.hh>
-#include <flecsi/run/control.hh>
+#include <flecsi/runtime.hh>
 
 enum class cp { advance };